# apps/api/src/main/resources/static/v1/openapi.yaml
openapi: 3.1.0

info:
  title: Fikaliako API
  version: 0.2.0
  summary: Geolocated food discovery for Madagascar
  description: |
    REST API of **Fikaliako**, the platform for discovering places to eat in
    Madagascar — restaurants, gargottes, cafés, street vendors — around you.

    ## Contract-first

    This document is the **manually managed source of truth** for the API
    (project book ch. 7.1, "one source contract"). It is written by hand,
    versioned with the code in
    `apps/api/src/main/resources/static/v1/openapi.yaml`, and served at
    `/v1/openapi.yaml`. Controllers are implemented **against** this contract,
    and the TypeScript client types for `apps/web` are generated **from** it.
    When behaviour and spec disagree, the spec wins: fix the code or amend the
    spec in the same change.

    Endpoints not yet implemented are marked with a (planned) note in their
    description; they define the target contract for the modules to come.

    ## Conventions

    - All endpoints are rooted at `/v1`.
    - Reads (`GET`) are public; writes require authentication (JWT, arriving
      with the accounts module — project book ch. 4.7).
    - Errors use RFC 9457 Problem Details (`application/problem+json`) and carry
      a `correlation_id`, also returned in the `X-Correlation-Id` header.
    - Prices are integers in **ariary** (`*_ar` fields).
    - Coordinates are WGS 84 (SRID 4326): `lat` ∈ [-90, 90], `lng` ∈ [-180, 180].
    - Times are local (Indian/Antananarivo), `HH:mm`; days of week are
      `0` = Monday … `6` = Sunday.
    - Lists are cursor-paginated: pass the returned `next_cursor` back as
      `?cursor=` to fetch the next page. A null/absent `next_cursor` is the end.
    - Fields that are null are omitted from responses.
  contact:
    name: Fikaliako

servers:
  - url: http://localhost:8080
    description: Local development

security: []

tags:
  - name: Health
    description: Liveness and service metadata.
  - name: Establishments
    description: Browse and read establishments (restaurants, gargottes, vendors…).
  - name: Search
    description: Geolocated and text discovery.
  - name: Reviews
    description: Five-criteria community reviews.
  - name: Account
    description: Authenticated user, favorites, contributions (planned).

paths:
  /v1/ping:
    get:
      tags: [Health]
      operationId: ping
      summary: Service liveness probe
      description: >
        Lightweight liveness check used by the web app's home page. Returns
        the service identity and a greeting. For operational health checks
        prefer the actuator endpoint (`/actuator/health`).
      responses:
        "200":
          description: Service is up.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PingResponse"
              example:
                service: fikaliako-api
                message: "Aiza no hisakafo androany ?"

  /v1/nearby:
    get:
      tags: [Search]
      operationId: nearby
      summary: Establishments near a point, cheapest first
      description: >
        The core discovery query (project book ch. 6.2): active establishments
        within `radius` metres of (`lat`, `lng`), ordered cheapest first then by
        distance. Uses `ST_DWithin` on the GiST index; target < 300 ms p95.
        The result set is bounded by `radius` and `limit`, so it is **not**
        cursor-paginated (`next_cursor` is always absent).
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number, format: double, minimum: -90, maximum: 90 }
        - name: lng
          in: query
          required: true
          schema: { type: number, format: double, minimum: -180, maximum: 180 }
        - name: radius
          in: query
          description: Search radius in metres (default 1000, max 5000).
          schema: { type: number, format: double, default: 1000, maximum: 5000 }
        - $ref: "#/components/parameters/TypeFilter"
        - $ref: "#/components/parameters/MinPriceFilter"
        - $ref: "#/components/parameters/MaxPriceFilter"
        - $ref: "#/components/parameters/CuisineFilter"
        - $ref: "#/components/parameters/PaymentFilter"
        - $ref: "#/components/parameters/BooleanFilters"
        - $ref: "#/components/parameters/OpenNowFilter"
        - name: limit
          in: query
          description: Max results (default 50, max 200 — the client clustering threshold).
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Ordered nearby results.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstablishmentPage"
        "400":
          $ref: "#/components/responses/BadRequest"

  /v1/search:
    get:
      tags: [Search]
      operationId: search
      summary: Typo-tolerant text search (planned)
      description: >
        (planned) Full-text search backed by Meilisearch synced from Postgres
        (project book ch. 7.2). Same combinable filters as `/v1/nearby`, plus a
        free-text `q`. Arrives with the search module.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/CursorParam"
      responses:
        "200":
          description: Matching establishments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstablishmentPage"

  /v1/establishments:
    get:
      tags: [Establishments]
      operationId: listEstablishments
      summary: Browse establishments
      description: >
        Active establishments, newest first, cursor-paginated. Filters combine
        (project book ch. 4.3).
      parameters:
        - $ref: "#/components/parameters/TypeFilter"
        - $ref: "#/components/parameters/MinPriceFilter"
        - $ref: "#/components/parameters/MaxPriceFilter"
        - $ref: "#/components/parameters/CuisineFilter"
        - $ref: "#/components/parameters/PaymentFilter"
        - $ref: "#/components/parameters/BooleanFilters"
        - $ref: "#/components/parameters/OpenNowFilter"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/CursorParam"
      responses:
        "200":
          description: A page of establishments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstablishmentPage"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [Establishments]
      operationId: proposeEstablishment
      summary: Propose a new establishment (planned)
      description: >
        (planned) Community contribution — anyone connected can propose a new
        establishment, positioned by dropping a pin (project book ch. 4.8). Goes
        through the moderation queue. Arrives with the community module.
      security:
        - bearerAuth: []
      responses:
        "202":
          description: Proposal accepted for moderation.

  /v1/establishments/{idOrSlug}:
    get:
      tags: [Establishments]
      operationId: getEstablishment
      summary: Establishment detail
      description: >
        Full detail for one establishment, resolvable by UUID or SEO slug (the
        slug path powers Google-indexable pages — project book ch. 11.3).
        `open_now` is computed server-side in Indian/Antananarivo.
      parameters:
        - name: idOrSlug
          in: path
          required: true
          description: The establishment UUID or its slug.
          schema: { type: string }
      responses:
        "200":
          description: The establishment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstablishmentDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  /v1/establishments/{establishmentId}/reviews:
    get:
      tags: [Reviews]
      operationId: listReviews
      summary: Reviews for an establishment
      description: Published reviews, newest first, cursor-paginated.
      parameters:
        - $ref: "#/components/parameters/EstablishmentId"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/CursorParam"
      responses:
        "200":
          description: A page of reviews.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewPage"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Reviews]
      operationId: createReview
      summary: Post a review (planned)
      description: >
        (planned) One review per user per establishment, scored 1–5 on five
        criteria (project book ch. 4.4). Arrives with the community module.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/EstablishmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewItem"

  /v1/users/me:
    get:
      tags: [Account]
      operationId: getCurrentUser
      summary: Current authenticated user (planned)
      description: "(planned) Arrives with the accounts module (project book ch. 4.7)."
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The authenticated user.

  /v1/users/me/favorites/{establishmentId}:
    put:
      tags: [Account]
      operationId: addFavorite
      summary: Add a favorite (planned)
      description: "(planned) Idempotent add to the user's favorites (project book ch. 4.6)."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/EstablishmentId"
      responses:
        "204":
          description: Favorited.
    delete:
      tags: [Account]
      operationId: removeFavorite
      summary: Remove a favorite (planned)
      description: "(planned) Idempotent removal from the user's favorites."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/EstablishmentId"
      responses:
        "204":
          description: Un-favorited.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        (planned) Arrives with the accounts module (project book ch. 7.3). Required
        on write operations (reviews, contributions, favorites).

  parameters:
    LimitParam:
      name: limit
      in: query
      description: Page size (default 50 for establishments, 20 for reviews; max 100).
      schema: { type: integer, minimum: 1, maximum: 100 }
    CursorParam:
      name: cursor
      in: query
      description: Opaque pagination cursor from a previous response's `next_cursor`.
      schema: { type: string }
    EstablishmentId:
      name: establishmentId
      in: path
      required: true
      description: Establishment UUID.
      schema: { type: string, format: uuid }
    TypeFilter:
      name: type
      in: query
      description: Filter by establishment type.
      schema:
        $ref: "#/components/schemas/EstablishmentType"
    MinPriceFilter:
      name: min_price
      in: query
      description: Minimum average price in ariary.
      schema: { type: integer, minimum: 0 }
    MaxPriceFilter:
      name: max_price
      in: query
      description: Maximum average price in ariary (the budget king filter, book ch. 3).
      schema: { type: integer, minimum: 0 }
    CuisineFilter:
      name: cuisine
      in: query
      description: Filter by cuisine referential code (e.g. `malagasy`, `grill`).
      schema: { type: string }
    PaymentFilter:
      name: payment
      in: query
      description: >
        A payment-method code (e.g. `mvola`) or the literal `mobile` to match any
        mobile-money operator (book ch. 4.3).
      schema: { type: string }
    BooleanFilters:
      name: filters
      in: query
      description: >
        Comma-separated boolean amenities that must all be true. One of:
        `delivery, parking, wifi, wheelchair_access, air_conditioning, terrace,
        family_friendly, romantic, student_friendly, scenic_view, open_24h`.
      schema: { type: string }
    OpenNowFilter:
      name: open_now
      in: query
      description: When true, keep only establishments open now (Indian/Antananarivo).
      schema: { type: boolean, default: false }

  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

  schemas:
    PingResponse:
      type: object
      description: Service identity and greeting.
      required: [service, message]
      properties:
        service:
          type: string
          description: Service identifier.
          examples: [fikaliako-api]
        message:
          type: string
          description: Human greeting, in Malagasy.
          examples: ["Aiza no hisakafo androany ?"]

    Problem:
      type: object
      description: RFC 9457 problem detail.
      required: [status, title]
      properties:
        type: { type: string, format: uri, default: "about:blank" }
        title: { type: string, examples: ["Not found"] }
        status: { type: integer, examples: [404] }
        detail: { type: string, examples: ["Establishment 'nope' not found."] }
        instance: { type: string, examples: ["/v1/establishments/nope"] }
        correlation_id:
          type: string
          description: Echoes the `X-Correlation-Id` response header.

    EstablishmentType:
      type: string
      description: Kind of eatery (project book ch. 6.1).
      enum:
        - restaurant
        - gargotte
        - cafe
        - snack
        - food_truck
        - street_vendor
        - pastry_shop
        - bar_restaurant
        - hotel_restaurant

    GeoPoint:
      type: object
      description: WGS 84 coordinate (SRID 4326).
      required: [lat, lng]
      properties:
        lat: { type: number, format: double, examples: [-18.9092] }
        lng: { type: number, format: double, examples: [47.5210] }

    ReferentialItem:
      type: object
      description: A referential entry with bilingual labels (FR + MG).
      required: [code, label_fr, label_mg]
      properties:
        code: { type: string, examples: [mvola] }
        label_fr: { type: string, examples: [MVola] }
        label_mg: { type: string, examples: [MVola] }

    Amenities:
      type: object
      description: The one-query boolean attributes (project book ch. 4.3 / 6.1).
      required:
        - delivery
        - parking
        - wifi
        - wheelchair_access
        - air_conditioning
        - terrace
        - family_friendly
        - romantic
        - student_friendly
        - scenic_view
        - open_24h
      properties:
        delivery: { type: boolean }
        parking: { type: boolean }
        wifi: { type: boolean }
        wheelchair_access: { type: boolean }
        air_conditioning: { type: boolean }
        terrace: { type: boolean }
        family_friendly: { type: boolean }
        romantic: { type: boolean }
        student_friendly: { type: boolean }
        scenic_view: { type: boolean }
        open_24h: { type: boolean }

    OpeningInterval:
      type: object
      description: One opening interval; days are 0 = Monday … 6 = Sunday.
      required: [day_of_week, opens_at, closes_at]
      properties:
        day_of_week: { type: integer, minimum: 0, maximum: 6 }
        opens_at: { type: string, examples: ["07:00"] }
        closes_at: { type: string, examples: ["21:00"] }

    RatingSummary:
      type: object
      description: >
        Aggregated ratings, recomputed nightly (project book ch. 4.5). Per-criterion
        and Bayesian fields are absent until an establishment has reviews.
      required: [count]
      properties:
        count: { type: integer, examples: [12] }
        avg_global: { type: number, examples: [4.5] }
        avg_quality: { type: number }
        avg_price: { type: number }
        avg_cleanliness: { type: number }
        avg_speed: { type: number }
        avg_welcome: { type: number }
        bayesian_note:
          type: number
          description: Dampened score used for rankings.

    EstablishmentSummary:
      type: object
      description: Card-level projection for lists and nearby results.
      required: [id, slug, name, type, position, verified, status, rating_count]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string, examples: [gargotte-chez-bao] }
        name: { type: string, examples: ["Gargotte Chez Bao"] }
        type:
          $ref: "#/components/schemas/EstablishmentType"
        position:
          $ref: "#/components/schemas/GeoPoint"
        avg_price_ar: { type: integer, examples: [2500] }
        verified: { type: boolean }
        status: { type: string, enum: [active, closed, pending] }
        rating_avg: { type: number, examples: [4.5] }
        rating_count: { type: integer, examples: [12] }
        distance_m:
          type: number
          description: Metres from the query point; present only on `/v1/nearby`.
          examples: [30.5]

    EstablishmentDetail:
      type: object
      description: Full establishment detail (project book ch. 8.2 — detail page).
      required:
        - id
        - slug
        - name
        - type
        - position
        - city
        - verified
        - status
        - open_now
        - amenities
        - opening_hours
        - payment_methods
        - cuisines
        - rating
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        type:
          $ref: "#/components/schemas/EstablishmentType"
        position:
          $ref: "#/components/schemas/GeoPoint"
        address: { type: string }
        district: { type: string, examples: [Analakely] }
        city: { type: string, examples: [Antananarivo] }
        phone: { type: string }
        whatsapp: { type: string }
        facebook_url: { type: string }
        website: { type: string }
        avg_price_ar: { type: integer }
        verified: { type: boolean }
        status: { type: string, enum: [active, closed, pending] }
        open_now:
          type: boolean
          description: Computed server-side in Indian/Antananarivo.
        amenities:
          $ref: "#/components/schemas/Amenities"
        opening_hours:
          type: array
          items:
            $ref: "#/components/schemas/OpeningInterval"
        payment_methods:
          type: array
          items:
            $ref: "#/components/schemas/ReferentialItem"
        cuisines:
          type: array
          items:
            $ref: "#/components/schemas/ReferentialItem"
        rating:
          $ref: "#/components/schemas/RatingSummary"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    ReviewItem:
      type: object
      description: A published review (project book ch. 4.4).
      required:
        - id
        - author_name
        - rating_quality
        - rating_price
        - rating_cleanliness
        - rating_speed
        - rating_welcome
        - global_note
        - created_at
      properties:
        id: { type: string, format: uuid }
        author_name: { type: string, examples: [Naina] }
        rating_quality: { type: integer, minimum: 1, maximum: 5 }
        rating_price: { type: integer, minimum: 1, maximum: 5 }
        rating_cleanliness: { type: integer, minimum: 1, maximum: 5 }
        rating_speed: { type: integer, minimum: 1, maximum: 5 }
        rating_welcome: { type: integer, minimum: 1, maximum: 5 }
        global_note:
          type: number
          description: Weighted mean with quality counted twice (weights sum to 6).
          examples: [4.67]
        comment: { type: string }
        created_at: { type: string, format: date-time }

    ReviewInput:
      type: object
      description: (planned) Payload to create a review.
      required:
        - rating_quality
        - rating_price
        - rating_cleanliness
        - rating_speed
        - rating_welcome
      properties:
        rating_quality: { type: integer, minimum: 1, maximum: 5 }
        rating_price: { type: integer, minimum: 1, maximum: 5 }
        rating_cleanliness: { type: integer, minimum: 1, maximum: 5 }
        rating_speed: { type: integer, minimum: 1, maximum: 5 }
        rating_welcome: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, maxLength: 2000 }

    EstablishmentPage:
      type: object
      description: A cursor-paginated page of establishment summaries.
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EstablishmentSummary"
        next_cursor:
          type: string
          description: Pass back as `?cursor=`; absent on the last page.

    ReviewPage:
      type: object
      description: A cursor-paginated page of reviews.
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReviewItem"
        next_cursor:
          type: string
          description: Pass back as `?cursor=`; absent on the last page.